corrected the pid file in test/manager.c \n was also added the debugfs subsystem to portsDB.c and corrected a value in a structure in info_acquire.h

+++ info_acquire.h
@@ -37,7 +37,7 @@ struct syscall_info_acquire {
 };
 
 struct db_info_acquire {
-	int how_many_ports;
+	long how_many_ports;
 	long how_many_inserts;
 	long how_many_removes;
 };
